@@ -59,11 +59,11 @@
         "react-mathjax-preview": "^0.1.16",
         "react-number-format": "^4.3.0",
         "react-range": "^1.3.0",
-        "react-redux": "^6.0.1",
+        "react-redux": "^7.2.0",
         "react-router": "^5.1.2",
         "react-router-config": "^5.0.1",
         "react-router-dom": "^5.0.1",
-        "react-scripts": "3.1.2",
+        "react-scripts": "^3.4.1",
         "react-select": "^3.0.4",
         "react-sortable-hoc": "^1.10.1",
         "react-split-pane": "^0.1.87",
@@ -100,7 +100,7 @@
         "enzyme": "^3.10.0",
         "enzyme-adapter-react-16": "^1.14.0",
         "enzyme-to-json": "^3.4.0",
-        "eslint": "^6.5.1",
+        "eslint": "^6.8.0",
         "eslint-config-prettier": "^6.4.0",
         "eslint-plugin-prettier": "^3.1.1",
         "husky": "^3.0.7",
