@@ -1,10 +1,10 @@
+import React, { useState, useEffect } from 'react';
 import { Container, Row, Col } from 'reactstrap';
-import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { connect } from 'react-redux';
 import { getUserInformationById } from 'network';
-import FilteredResourcesList from './FilteredResourcesList';
+import Items from './Items';
 import NotFound from 'components/StaticPages/NotFound';
+import { useSelector } from 'react-redux';
 
 /*
 const StyledGravatar = styled(Gravatar)`
@@ -41,119 +41,104 @@ const StyledActivity = styled.div`
 `;
 */
 
-class UserProfile extends Component {
-    state = {
-        displayName: '',
-        notFound: false
-    };
+const UserProfile = props => {
+    const [displayName, setDisplayName] = useState('');
+    const [notFound, setNotFound] = useState(false);
+    const userId = props.match.params.userId;
+    const currentUserId = useSelector(state => state.auth.user?.id);
 
-    componentDidMount() {
-        this.getUserInformation();
+    useEffect(() => {
+        const getUserInformation = async () => {
+            setNotFound(false);
+
+            try {
+                const userData = await getUserInformationById(userId);
+                const { display_name: displayName } = userData;
+
+                setDisplayName(displayName);
+            } catch (e) {
+                setNotFound(true);
+            }
+        };
+
+        getUserInformation();
+    }, [userId]);
+
+    if (notFound) {
+        return <NotFound />;
     }
 
-    getUserInformation = async () => {
-        this.setState({
-            notFound: false
-        });
-
-        try {
-            const userData = await getUserInformationById(this.props.match.params.userId);
-
-            this.setState({
-                displayName: userData.display_name,
-                email: userData.email
-            });
-        } catch (e) {
-            this.setState({
-                notFound: true
-            });
-        }
-    };
-
-    render() {
-        if (this.state.notFound) {
-            return <NotFound />;
-        }
-
-        return (
-            <>
-                <Container className="p-0">
-                    <h1 className="h4 mt-4 mb-4">Profile of {this.state.displayName}</h1>
-                </Container>
-                {/*<Container className="box pt-4 pb-3 pl-5 pr-5">
-                    <Row>
-                        <div className="col-1 text-center">
-                            <StyledGravatar className="rounded-circle" email="example@example.com" size={76} id="TooltipExample" />{' '}
-                            TODO: Replace with hash from email (should be returned by the backend) 
-                        </div>
-                        <div className="col-11 pl-4">
-                            <h2 className="h5">{this.state.displayName}</h2>
-                            TODO: support more information for users, like organization and short bio
-                            <p>
-                            <b className={'d-block'}>Organization</b>
-                            L3S Research Center
-                        </p>
-
+    return (
+        <>
+            <Container className="p-0">
+                <h1 className="h4 mt-4 mb-4">Profile of {displayName}</h1>
+            </Container>
+            {/*<Container className="box pt-4 pb-3 pl-5 pr-5">
+                <Row>
+                    <div className="col-1 text-center">
+                        <StyledGravatar className="rounded-circle" email="example@example.com" size={76} id="TooltipExample" />{' '}
+                        TODO: Replace with hash from email (should be returned by the backend) 
+                    </div>
+                    <div className="col-11 pl-4">
+                        <h2 className="h5">{this.state.displayName}</h2>
+                        TODO: support more information for users, like organization and short bio
                         <p>
-                            <b className={'d-block'}>Bio </b>
-                            Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna
-                            aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
-                        </p>
-                        </div>
-                    </Row>
-                </Container>*/}
-                <Container className="box mt-4 pt-4 pb-3 pl-5 pr-5">
-                    <Row>
-                        <Col md={6} sm={12} style={{ display: 'flex', flexDirection: 'column' }}>
-                            <h5 className="mb-4">Added papers</h5>
-                            <FilteredResourcesList
-                                filterLabel={'papers'}
-                                filterClass={process.env.REACT_APP_CLASSES_PAPER}
-                                userId={this.props.match.params.userId}
-                            />
-                        </Col>
-                        <Col md={6} sm={12} style={{ display: 'flex', flexDirection: 'column' }}>
-                            <h5 className="mb-4">Published comparisons</h5>
-                            <FilteredResourcesList
-                                filterLabel={'comparisons'}
-                                filterClass={process.env.REACT_APP_CLASSES_COMPARISON}
-                                userId={this.props.match.params.userId}
-                            />
-                        </Col>
-                    </Row>
-                </Container>
-                {/*
-                TODO: support for activity feed
-                <Container className="box mt-4 pt-4 pb-3 pl-5 pr-5">
-                <h5 className="mb-4">Activity feed</h5>
-                <StyledActivity className="pl-3 pb-3">
-                    <div className={'time'}>16 JULY 2019</div>
-                    <div>
-                        John Doe updated resource <Link to={'/'}>IoT research directions</Link>
-                    </div>
-                </StyledActivity>
-                <StyledActivity className="pl-3 pb-3">
-                    <div className={'time'}>10 JULY 2019</div>
-                    <div>
-                        John Doe updated resource <Link to={'/'}>IoT research directions</Link>
-                    </div>
-                    <div>
-                        John Doe commented on predicate <Link to={'/'}>Has Problem</Link>
-                    </div>
-                </StyledActivity>
-                <StyledActivity className="pl-3 pb-3">
-                    <div className={'time'}>5 JULY 2019</div>
-                    <div>John Doe joined ORKG, welcome!</div>
-                </StyledActivity>
-                </Container>*/}
-            </>
-        );
-    }
-}
+                        <b className={'d-block'}>Organization</b>
+                        L3S Research Center
+                    </p>
 
-const mapStateToProps = state => ({});
-
-const mapDispatchToProps = dispatch => ({});
+                    <p>
+                        <b className={'d-block'}>Bio </b>
+                        Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna
+                        aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
+                    </p>
+                    </div>
+                </Row>
+            </Container>*/}
+            <Container className="box mt-4 pt-4 pb-3 pl-5 pr-5">
+                <Row>
+                    <Col md={6} sm={12} style={{ display: 'flex', flexDirection: 'column' }}>
+                        <h5 className="mb-4">Added papers</h5>
+                        <Items
+                            filterLabel={'papers'}
+                            filterClass={process.env.REACT_APP_CLASSES_PAPER}
+                            userId={userId}
+                            showDelete={userId === currentUserId}
+                        />
+                    </Col>
+                    <Col md={6} sm={12} style={{ display: 'flex', flexDirection: 'column' }}>
+                        <h5 className="mb-4">Published comparisons</h5>
+                        <Items filterLabel={'comparisons'} filterClass={process.env.REACT_APP_CLASSES_COMPARISON} userId={userId} />
+                    </Col>
+                </Row>
+            </Container>
+            {/*
+            TODO: support for activity feed
+            <Container className="box mt-4 pt-4 pb-3 pl-5 pr-5">
+            <h5 className="mb-4">Activity feed</h5>
+            <StyledActivity className="pl-3 pb-3">
+                <div className={'time'}>16 JULY 2019</div>
+                <div>
+                    John Doe updated resource <Link to={'/'}>IoT research directions</Link>
+                </div>
+            </StyledActivity>
+            <StyledActivity className="pl-3 pb-3">
+                <div className={'time'}>10 JULY 2019</div>
+                <div>
+                    John Doe updated resource <Link to={'/'}>IoT research directions</Link>
+                </div>
+                <div>
+                    John Doe commented on predicate <Link to={'/'}>Has Problem</Link>
+                </div>
+            </StyledActivity>
+            <StyledActivity className="pl-3 pb-3">
+                <div className={'time'}>5 JULY 2019</div>
+                <div>John Doe joined ORKG, welcome!</div>
+            </StyledActivity>
+            </Container>*/}
+        </>
+    );
+};
 
 UserProfile.propTypes = {
     match: PropTypes.shape({
@@ -163,7 +148,4 @@ UserProfile.propTypes = {
     }).isRequired
 };
 
-export default connect(
-    mapStateToProps,
-    mapDispatchToProps
-)(UserProfile);
+export default UserProfile;
